Implementación de Registro de Usuario y Módulo de Beneficiarios

+++ src/main/java/org/walrex/application/port/input/DeleteBeneficiaryAccountUseCase.java
@@ -3,5 +3,5 @@ package org.walrex.application.port.input;
 import io.smallrye.mutiny.Uni;
 
 public interface DeleteBeneficiaryAccountUseCase {
-    Uni<Void> delete(Integer id);
+    Uni<Void> delete(Long id);
 }
