+++ src/main/java/org/walrex/domain/exception/BeneficiaryAccountNotFoundException.java
@@ -4,7 +4,7 @@ import jakarta.ws.rs.WebApplicationException;
 import jakarta.ws.rs.core.Response;
 
 public class BeneficiaryAccountNotFoundException extends WebApplicationException {
-    public BeneficiaryAccountNotFoundException(Integer id) {
+    public BeneficiaryAccountNotFoundException(Long id) {
         super("Beneficiary Account with ID " + id + " not found.", Response.Status.NOT_FOUND);
     }
 }
